requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
cffi==1.15.0
cryptography==35.0.0
distlib==0.3.3
Django==3.2.8
django-rest-knox==4.1.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
filelock==3.3.1
gunicorn==20.1.0
pipenv==2021.5.29
platformdirs==2.4.0
pycparser==2.20
pymongo==3.12.1
pytz==2021.3
six==1.16.0
sqlparse==0.2.4
virtualenv==20.9.0
virtualenv-clone==0.5.7
